feat(#73): small refactoring for paths
