+++ apps/server/src/app/auth/utils/jwt-auth.strategy.ts
@@ -2,7 +2,7 @@ import { ExtractJwt, Strategy } from 'passport-jwt'
 import { PassportStrategy } from '@nestjs/passport'
 import { Injectable, UnauthorizedException } from '@nestjs/common'
 import { ConfigService } from '@nestjs/config'
-import { User } from '@users/entities/user';
+import { User } from '@app/users';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
