+++ apps/server/src/app/roles/index.ts
@@ -1,0 +1,3 @@
+export * from './constants/actions'
+export * from './providers/ability.factory'
+export * from './roles.module'
