+++ apps/server/src/app/users/index.ts
@@ -1,3 +1,3 @@
 export * from "./users.module"
-export * from "./users.service"
+export * from "./providers/users.service"
 export * from "./entities/user";
